enable embedding search (#164)
## Purpose
<!-- Describe the intention of the changes being proposed. What problem
does it solve or functionality does it add? -->
* ...

## Does this introduce a breaking change?
<!-- Mark one with an "x". -->
```
[ ] Yes
[ ] No
```
fix #120

This PR enables option for three search mode
- Text
- Vector
- Hybrid

When `Vector` or `Hybrid` mode is enabled, vector search will be enabled
when searching document from index.

<img width="443" alt="image"
src="https://github.com/Azure-Samples/azure-search-openai-demo-csharp/assets/16876986/6326ce20-bed7-49f5-aae0-e62f2505dcdd">


## Pull Request Type
What kind of change does this Pull Request introduce?

<!-- Please check the one that applies to this PR using "x". -->
```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Documentation content changes
[ ] Other... Please describe:
```

## How to Test
*  Get the code

```
git clone [repo-address]
cd [repo-name]
git checkout [branch-name]
npm install
```

* Test the code
<!-- Add steps to run the tests suite and/or manually test -->
```
```

## What to Check
Verify that the following are valid
* ...

## Other Information
<!-- Add any other helpful information that may be needed here. -->

+++ app/prepdocs/PrepareDocs/Program.cs
@@ -144,6 +144,10 @@ static async ValueTask RemoveFromIndexAsync(
             });
         }
 
+        if (documentsToDelete.Count == 0)
+        {
+            break;
+        }
         Response<IndexDocumentsResult> deleteResponse =
             await searchClient.DeleteDocumentsAsync(documentsToDelete);
 
@@ -176,15 +180,30 @@ static async ValueTask CreateSearchIndexAsync(AppOptions options)
         }
     }
 
+    string vectorSearchConfigName = "my-vector-config";
+
     var index = new SearchIndex(options.SearchIndexName)
     {
+        VectorSearch = new()
+        {
+            AlgorithmConfigurations =
+                {
+                    new HnswVectorSearchAlgorithmConfiguration(vectorSearchConfigName)
+                }
+        },
         Fields =
         {
             new SimpleField("id", SearchFieldDataType.String) { IsKey = true },
             new SearchableField("content") { AnalyzerName = "en.microsoft" },
             new SimpleField("category", SearchFieldDataType.String) { IsFacetable = true },
             new SimpleField("sourcepage", SearchFieldDataType.String) { IsFacetable = true },
-            new SimpleField("sourcefile", SearchFieldDataType.String) { IsFacetable = true }
+            new SimpleField("sourcefile", SearchFieldDataType.String) { IsFacetable = true },
+            new SearchField("embedding", SearchFieldDataType.Collection(SearchFieldDataType.Single))
+            {
+                VectorSearchDimensions = 1536,
+                IsSearchable = true,
+                VectorSearchConfiguration = vectorSearchConfigName,
+            }
         },
         SemanticSettings = new SemanticSettings
         {
@@ -502,11 +521,14 @@ static async ValueTask IndexSectionsAsync(
     }
 
     var searchClient = await GetSearchClientAsync(options);
+    var openAIClient = await GetAzureOpenAIClientAsync(options);
 
     var iteration = 0;
     var batch = new IndexDocumentsBatch<SearchDocument>();
     foreach (var section in sections)
     {
+        var embeddings = await openAIClient.GetEmbeddingsAsync(options.EmbeddingModelName, new Azure.AI.OpenAI.EmbeddingsOptions(section.Content.Replace('\r', ' ')));
+        var embedding = embeddings.Value.Data.FirstOrDefault()?.Embedding.ToArray() ?? new float[0];
         batch.Actions.Add(new IndexDocumentsAction<SearchDocument>(
             IndexActionType.MergeOrUpload,
             new SearchDocument
@@ -515,7 +537,8 @@ static async ValueTask IndexSectionsAsync(
                 ["content"] = section.Content,
                 ["category"] = section.Category,
                 ["sourcepage"] = section.SourcePage,
-                ["sourcefile"] = section.SourceFile
+                ["sourcefile"] = section.SourceFile,
+                ["embedding"] = embedding,
             }));
 
         iteration++;
